+++ pkg/ddc/base/runtime_helper.go
@@ -53,10 +53,10 @@ func init() {
 }
 
 // GetTemplateToInjectForFuse gets template for fuse injection
-func (info *RuntimeInfo) GetTemplateToInjectForFuse(pvcName string, option common.FuseSidecarInjectOption) (template *common.FuseInjectionTemplate, err error) {
+func (info *RuntimeInfo) GetTemplateToInjectForFuse(pvcName string, pvcNamespace string, option common.FuseSidecarInjectOption) (template *common.FuseInjectionTemplate, err error) {
 	if utils.IsTimeTrackerDebugEnabled() {
 		defer utils.TimeTrack(time.Now(), "RuntimeInfo.GetTemplateToInjectForFuse",
-			"pvc.name", pvcName, "pvc.namespace", info.GetNamespace())
+			"pvc.name", pvcName, "pvc.namespace", pvcNamespace)
 	}
 	// TODO: create fuse container
 	ds, err := info.getFuseDaemonset()
@@ -64,7 +64,8 @@ func (info *RuntimeInfo) GetTemplateToInjectForFuse(pvcName string, option commo
 		return template, err
 	}
 
-	dataset, err := utils.GetDataset(info.client, info.name, info.namespace)
+	// Note: get the pvc corresponding dataset
+	dataset, err := utils.GetDataset(info.client, pvcName, pvcNamespace)
 	if err != nil {
 		return template, err
 	}
@@ -119,9 +120,10 @@ func (info *RuntimeInfo) GetTemplateToInjectForFuse(pvcName string, option commo
 		mountPathInContainer = volumeMountInContainer.MountPath
 	}
 
+	// Fluid assumes pvc name is the same as runtime's name
 	gen := poststart.NewGenerator(types.NamespacedName{
-		Name:      info.name,
-		Namespace: info.namespace,
+		Name:      pvcName,
+		Namespace: pvcNamespace,
 	}, mountPathInContainer, mountType, option)
 	cm := gen.BuildConfigmap(ownerReference)
 	found, err := kubeclient.IsConfigMapExist(info.client, cm.Name, cm.Namespace)