+++ pkg/ddc/eac/port_parser_test.go
@@ -1,0 +1,88 @@
+package eac
+
+import (
+	"github.com/fluid-cloudnative/fluid/api/v1alpha1"
+	"github.com/fluid-cloudnative/fluid/pkg/common"
+	"github.com/fluid-cloudnative/fluid/pkg/utils/fake"
+	v1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"testing"
+)
+
+var valuesConfigMapData = `
+`
+
+func Test_parsePortsFromConfigMap(t *testing.T) {
+}
+
+func TestGetReservedPorts(t *testing.T) {
+	configMap := &v1.ConfigMap{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "hbase-eac-values",
+			Namespace: "fluid",
+		},
+		Data: map[string]string{
+			"data": valuesConfigMapData,
+		},
+	}
+	dataSets := []*v1alpha1.Dataset{
+		{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "hbase",
+				Namespace: "fluid",
+			},
+			Status: v1alpha1.DatasetStatus{
+				Runtimes: []v1alpha1.Runtime{
+					{
+						Name:      "hbase",
+						Namespace: "fluid",
+						Type:      common.EACRuntimeType,
+					},
+				},
+			},
+		},
+		{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "no-runtime",
+				Namespace: "fluid",
+			},
+		},
+		{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "runtime-type",
+				Namespace: "fluid",
+			},
+			Status: v1alpha1.DatasetStatus{
+				Runtimes: []v1alpha1.Runtime{
+					{
+						Type: "not-eac",
+					},
+				},
+			},
+		},
+		{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "no-map",
+				Namespace: "fluid",
+			},
+			Status: v1alpha1.DatasetStatus{
+				Runtimes: []v1alpha1.Runtime{
+					{
+						Type: common.EACRuntimeType,
+					},
+				},
+			},
+		},
+	}
+	var runtimeObjs []runtime.Object
+	runtimeObjs = append(runtimeObjs, configMap)
+	for _, dataSet := range dataSets {
+		runtimeObjs = append(runtimeObjs, dataSet.DeepCopy())
+	}
+	fakeClient := fake.NewFakeClientWithScheme(testScheme, runtimeObjs...)
+	_, err := GetReservedPorts(fakeClient)
+	if err != nil {
+		t.Errorf("GetReservedPorts failed.")
+	}
+}