+++ pkg/ddc/factory.go
@@ -16,6 +16,7 @@ import (
 	"github.com/fluid-cloudnative/fluid/pkg/controllers/deploy"
 	"github.com/fluid-cloudnative/fluid/pkg/ddc/alluxio"
 	"github.com/fluid-cloudnative/fluid/pkg/ddc/base"
+	"github.com/fluid-cloudnative/fluid/pkg/ddc/eac"
 	"github.com/fluid-cloudnative/fluid/pkg/ddc/goosefs"
 	"github.com/fluid-cloudnative/fluid/pkg/ddc/jindo"
 	"github.com/fluid-cloudnative/fluid/pkg/ddc/jindofsx"
@@ -39,6 +40,7 @@ func init() {
 		"goosefs":  goosefs.Build,
 		"juicefs":  juicefs.Build,
 		"thin":     thin.Build,
+		"eac":      eac.Build,
 	}
 
 	deploy.SetPrecheckFunc(map[string]deploy.CheckFunc{