+++ pkg/ddc/thin/engine_test.go
@@ -26,6 +26,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 	"testing"
 )
 
@@ -140,5 +141,187 @@ func TestBuild(t *testing.T) {
 	gott, err := Build("testId", errrCtx)
 	if err == nil {
 		t.Errorf("expect err, but no err got %v", gott)
+	}
+}
+
+func TestBuildReferenceDatasetEngine(t *testing.T) {
+	var namespace = v1.Namespace{
+		ObjectMeta: metav1.ObjectMeta{
+			Name: "fluid",
+		},
+	}
+	testObjs := []runtime.Object{}
+	testObjs = append(testObjs, namespace.DeepCopy())
+
+	var dataset = datav1alpha1.Dataset{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "done",
+			Namespace: "big-data",
+		},
+		Status: datav1alpha1.DatasetStatus{
+			Runtimes: []datav1alpha1.Runtime{
+				{
+					Name:      "done",
+					Namespace: "big-data",
+					Type:      common.AlluxioRuntime,
+				},
+			},
+		},
+	}
+	var runtime = datav1alpha1.AlluxioRuntime{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "done",
+			Namespace: "big-data",
+		},
+	}
+
+	var refRuntime = datav1alpha1.ThinRuntime{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "hbase",
+			Namespace: "fluid",
+		},
+	}
+	var refDataset = datav1alpha1.Dataset{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "hbase",
+			Namespace: "fluid",
+		},
+		Spec: datav1alpha1.DatasetSpec{
+			Mounts: []datav1alpha1.Mount{
+				{
+					MountPoint: "dataset://big-data/done",
+				},
+			},
+		},
+	}
+
+	testObjs = append(testObjs, &dataset, &refDataset)
+
+	testObjs = append(testObjs, &runtime, &refRuntime)
+	client := fake.NewFakeClientWithScheme(testScheme, testObjs...)
+
+	var ctx = cruntime.ReconcileRequestContext{
+		NamespacedName: types.NamespacedName{
+			Name:      "hbase",
+			Namespace: "fluid",
+		},
+		Client:      client,
+		Log:         fake.NullLogger(),
+		RuntimeType: "thin",
+		Runtime:     &refRuntime,
+	}
+
+	engine, err := Build("testId", ctx)
+	if err != nil || engine == nil {
+		t.Errorf("fail to exec the build function with the eror %v", err)
+	}
+
+	var errCtx = cruntime.ReconcileRequestContext{
+		NamespacedName: types.NamespacedName{
+			Name:      "hbase",
+			Namespace: "fluid",
+		},
+		Client:      client,
+		Log:         fake.NullLogger(),
+		RuntimeType: "thin",
+		Runtime:     &runtime,
+	}
+
+	got, err := Build("testId", errCtx)
+	if err == nil {
+		t.Errorf("expect err, but no err got %v", got)
+	}
+}
+
+func TestIsReferenceDatasetRuntime(t *testing.T) {
+	tests := []struct {
+		name    string
+		dataset *datav1alpha1.Dataset
+		runtime *datav1alpha1.ThinRuntime
+		client  client.Client
+		want    bool
+		wantErr bool
+	}{
+		{
+			name: "ref-dataset",
+			dataset: &datav1alpha1.Dataset{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "hbase",
+					Namespace: "fluid",
+				},
+				Spec: datav1alpha1.DatasetSpec{
+					Mounts: []datav1alpha1.Mount{
+						{
+							MountPoint: "dataset://test/test",
+						},
+					},
+				},
+			},
+			runtime: &datav1alpha1.ThinRuntime{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "hbase",
+					Namespace: "fluid",
+				},
+				Spec: datav1alpha1.ThinRuntimeSpec{},
+			},
+			want:    true,
+			wantErr: false,
+		},
+		{
+			name: "not-ref-dataset",
+			dataset: &datav1alpha1.Dataset{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "hbase",
+					Namespace: "fluid",
+				},
+			},
+			runtime: &datav1alpha1.ThinRuntime{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "hbase",
+					Namespace: "fluid",
+				},
+				Spec: datav1alpha1.ThinRuntimeSpec{
+					ThinRuntimeProfileName: "",
+				},
+			},
+			want:    false,
+			wantErr: false,
+		},
+		{
+			name: "dataset-not-exist",
+			dataset: &datav1alpha1.Dataset{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "hbase-no-use",
+					Namespace: "fluid",
+				},
+			},
+			runtime: &datav1alpha1.ThinRuntime{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "hbase",
+					Namespace: "fluid",
+				},
+				Spec: datav1alpha1.ThinRuntimeSpec{
+					ThinRuntimeProfileName: "1",
+				},
+			},
+			want:    false,
+			wantErr: true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			fakeClient := fake.NewFakeClientWithScheme(testScheme, tt.dataset, tt.runtime)
+
+			isRef, err := isReferenceDatasetRuntime(fakeClient, tt.runtime)
+
+			if (err != nil) != tt.wantErr {
+				t.Errorf("expect has error %t, but get error %v", tt.wantErr, err)
+				return
+			}
+
+			if isRef != tt.want {
+				t.Errorf(" expect is ref dataset %t, but get %t", tt.want, err)
+			}
+		})
 	}
 }